provision/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile ./requirements.in
#
alembic==1.10.3
    # via -r ./requirements.in
anyio==3.6.2
    # via
    #   httpcore
    #   starlite
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
click==8.1.3
    # via uvicorn
dnspython==2.3.0
    # via email-validator
email-validator==2.0.0.post1
    # via pydantic-openapi-schema
faker==18.4.0
    # via pydantic-factories
fast-query-parsers==0.4.0
    # via starlite
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.17.0
    # via httpx
httpx==0.24.0
    # via starlite
idna==3.4
    # via
    #   anyio
    #   email-validator
    #   httpx
jinja2==3.1.2
    # via -r ./requirements.in
mako==1.2.4
    # via
    #   alembic
    #   starlite
markupsafe==2.1.2
    # via
    #   jinja2
    #   mako
msgspec==0.14.1
    # via starlite
multidict==6.0.4
    # via starlite
psycopg2==2.9.6
    # via -r ./requirements.in
pydantic==1.10.7
    # via
    #   pydantic-factories
    #   pydantic-openapi-schema
    #   starlite
pydantic-factories==1.17.3
    # via starlite
pydantic-openapi-schema==1.5.1
    # via starlite
pymysql==1.0.3
    # via -r ./requirements.in
python-dateutil==2.8.2
    # via faker
pyyaml==6.0
    # via starlite
six==1.16.0
    # via python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlalchemy==2.0.9
    # via
    #   -r ./requirements.in
    #   alembic
starlite==1.51.10
    # via -r ./requirements.in
typing-extensions==4.5.0
    # via
    #   alembic
    #   pydantic
    #   pydantic-factories
    #   sqlalchemy
    #   starlite
uvicorn==0.21.1
    # via -r ./requirements.in
